fix(host): separate styles, clean code
