+++ src/screens/explore/HostProfile/HostProfile.tsx
@@ -1,14 +1,6 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { StatusBar } from 'expo-status-bar';
 import { useCallback, useEffect } from 'react';
-import {
-  RefreshControl,
-  SafeAreaView,
-  ScrollView,
-  StyleSheet,
-  TouchableOpacity,
-  View,
-} from 'react-native';
+import { RefreshControl, ScrollView, TouchableOpacity, View } from 'react-native';
 import { useSelector } from 'react-redux';
 import { Icon, Loader } from 'src/components';
 import { ScreenTemplate } from 'src/components/templates';
@@ -16,14 +8,14 @@ import { RootStackParamList } from 'src/navigation';
 import { useAppDispatch } from 'src/store';
 import { getColors, getHostProfile, getHostProfileLoader } from 'src/store/selectors';
 import { AsyncThunks } from 'src/store/thunks';
-import { WHITE_100, LEVEL_1 } from 'src/styles';
 import { IconName } from 'src/types';
 
-import HostAbout from './components/HostAbout';
-import HostListings from './components/HostListings';
-import HostMainCard from './components/HostMainCard';
-import HostReviews from './components/HostReviews';
-import HostVerifiedInfo from './components/HostVerifiedInfo';
+import styles from './HostProfile.styles';
+import HostAbout from './components/HostAbout/HostAbout';
+import HostListings from './components/HostListings/HostListings';
+import HostMainCard from './components/HostMainCard/HostMainCard';
+import HostReviews from './components/HostReviews/HostReviews';
+import HostVerifiedInfo from './components/HostVerifiedInfo/HostVerifiedInfo';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'HostProfile'>;
 
@@ -48,11 +40,7 @@ const HostProfile = ({ route, navigation }: Props) => {
   }, [fetchHostProfile]);
 
   return (
-    <ScreenTemplate>
-      <SafeAreaView>
-        <StatusBar />
-      </SafeAreaView>
-
+    <ScreenTemplate headerShown={false}>
       <ScrollView
         refreshControl={
           <RefreshControl
@@ -84,31 +72,5 @@ const HostProfile = ({ route, navigation }: Props) => {
     </ScreenTemplate>
   );
 };
-
-const styles = StyleSheet.create({
-  screen: {
-    position: 'relative',
-  },
-  header: {
-    width: '100%',
-    height: 80,
-    padding: 16,
-    position: 'absolute',
-    zIndex: 1,
-    top: 32,
-    left: 0,
-    right: 0,
-  },
-  icon: {
-    flexDirection: 'row',
-    borderRadius: 100,
-    width: 40,
-    height: 40,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: WHITE_100,
-    ...LEVEL_1,
-  },
-});
 
 export default HostProfile;
