+++ src/screens/explore/HostProfile/components/HostAbout/HostAbout.tsx
@@ -1,11 +1,12 @@
 import { MaterialIcons } from '@expo/vector-icons';
 import dayjs from 'dayjs';
-import { View, StyleSheet } from 'react-native';
+import { View } from 'react-native';
 import { useSelector } from 'react-redux';
 import { Text } from 'src/components';
 import { getColors } from 'src/store/selectors';
-import { paragraph1 } from 'src/styles';
 import { HostProfile } from 'src/types';
+
+import styles from './HostAbout.styles';
 
 interface Props {
   host: HostProfile;
@@ -37,36 +38,5 @@ const HostAbout = ({ host }: Props) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'column',
-    padding: 16,
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  infoContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    marginBottom: 16,
-  },
-  infoItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginRight: 16,
-    marginBottom: 8,
-  },
-  infoText: {
-    ...paragraph1,
-    marginLeft: 8,
-  },
-  description: {
-    ...paragraph1,
-    fontWeight: '500',
-  },
-});
 
 export default HostAbout;
