+++ src/screens/explore/HostProfile/components/HostListings/HostListings.tsx
@@ -1,12 +1,13 @@
 import { useState } from 'react';
-import { View, TouchableOpacity, StyleSheet, Image } from 'react-native';
+import { View, TouchableOpacity, Image } from 'react-native';
 import { AirbnbRating } from 'react-native-ratings';
 import Swiper from 'react-native-swiper';
 import { useSelector } from 'react-redux';
 import { Text } from 'src/components';
 import { getColors } from 'src/store/selectors';
-import { GREY_200, subtitle1 } from 'src/styles';
 import { HostProfile } from 'src/types';
+
+import styles from './HostListings.styles';
 
 interface Props {
   host: HostProfile;
@@ -68,57 +69,5 @@ const HostListings = ({ host, navigation }: Props) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    paddingVertical: 20,
-    marginBottom: 50,
-  },
-  titleContainer: {
-    marginHorizontal: 10,
-  },
-  title: {
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  wrapper: {
-    height: 380,
-  },
-  slide: {
-    flex: 1,
-    padding: 24,
-  },
-  dotContainer: {
-    bottom: -20,
-  },
-  imageContainer: {
-    overflow: 'hidden',
-    borderRadius: 15,
-    borderWidth: 1,
-    borderColor: GREY_200,
-  },
-  accommodationImage: {
-    padding: 0,
-    width: '100%',
-    aspectRatio: 4 / 3,
-    objectFit: 'cover',
-  },
-  infoContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    flex: 1,
-    marginTop: 10,
-  },
-  titleText: {
-    ...subtitle1,
-    marginBottom: 5,
-    overflow: 'hidden',
-  },
-  rating: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-});
 
 export default HostListings;
