+++ src/screens/explore/HostProfile/components/HostMainCard/HostMainCard.tsx
@@ -1,11 +1,12 @@
 import { Ionicons } from '@expo/vector-icons';
-import { Image, StyleSheet, View } from 'react-native';
+import { Image, View } from 'react-native';
 import { useSelector } from 'react-redux';
 import { Text } from 'src/components';
 import { getColors } from 'src/store/selectors';
-import { PRIMARY_BLUE_200, title1, title3, GREY_300, subtitle1, WHITE_100 } from 'src/styles';
+import { PRIMARY_BLUE_200 } from 'src/styles';
 import { HostProfile } from 'src/types';
-import { PADDING_HORIZONTAL_28 } from 'src/utils';
+
+import styles from './HostMainCard.styles';
 
 interface Props {
   host: HostProfile;
@@ -55,64 +56,5 @@ const HostProfileCard = ({ host }: Props) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    padding: PADDING_HORIZONTAL_28,
-  },
-  userContainer: {
-    marginTop: 36,
-    flex: 6,
-    marginBottom: 16,
-    alignItems: 'center',
-  },
-  avatarContainer: {
-    position: 'relative',
-    marginBottom: 8,
-  },
-  profileImage: {
-    width: 150,
-    height: 150,
-    borderRadius: 100,
-    borderColor: WHITE_100,
-    borderWidth: 1,
-  },
-  verifiedIcon: {
-    position: 'absolute',
-    bottom: -2,
-    right: 0,
-  },
-  userInfo: {
-    flex: 1,
-  },
-  name: {
-    ...title1,
-  },
-  subtitle: {
-    ...title3,
-    fontWeight: 'normal',
-    color: GREY_300,
-  },
-  statsContainer: {
-    flex: 4,
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  statsItem: {
-    alignItems: 'center',
-  },
-  statsCount: {
-    ...subtitle1,
-    fontWeight: '600',
-  },
-  statsLabel: {
-    fontSize: 12,
-    color: GREY_300,
-  },
-  rating: {
-    paddingLeft: 2,
-  },
-});
 
 export default HostProfileCard;
