+++ src/screens/explore/HostProfile/components/HostReviews/HostReviews.tsx
@@ -1,11 +1,12 @@
 import dayjs from 'dayjs';
-import { View, Image, StyleSheet } from 'react-native';
+import { View, Image } from 'react-native';
 import Swiper from 'react-native-swiper';
 import { useSelector } from 'react-redux';
 import { Text } from 'src/components';
 import { getColors } from 'src/store/selectors';
-import { paragraph2, paragraph1, subtitle1, GREY_200 } from 'src/styles';
 import { HostProfile } from 'src/types';
+
+import styles from './HostReviews.styles';
 
 interface Props {
   host: HostProfile;
@@ -49,65 +50,5 @@ const HostReviews = ({ host }: Props) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    paddingVertical: 20,
-    marginBottom: 20,
-  },
-  wrapper: {
-    height: 210,
-  },
-  slide: {
-    flex: 1,
-    padding: 24,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  dotContainer: {
-    bottom: -20,
-  },
-  titleContainer: {
-    marginBottom: 20,
-    marginHorizontal: 20,
-  },
-  title: {
-    ...subtitle1,
-    fontWeight: 'bold',
-  },
-  reviewContainer: {
-    width: '100%',
-    borderRadius: 15,
-    borderWidth: 1,
-    borderColor: GREY_200,
-    padding: 20,
-  },
-  reviewContent: {
-    marginBottom: 15,
-  },
-  reviewText: {
-    ...paragraph1,
-  },
-  userInfoContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  userImage: {
-    width: 50,
-    height: 50,
-    borderRadius: 25,
-    marginRight: 10,
-  },
-  userInfoText: {
-    flexDirection: 'column',
-  },
-  userName: {
-    ...paragraph1,
-    fontWeight: 'bold',
-  },
-  joinedDate: {
-    ...paragraph2,
-  },
-});
 
 export default HostReviews;
